+++ docs/release-notes.md
@@ -22,6 +22,7 @@
 * Added a way for mods to get a list of all resolved mods (including ones which may not have been loaded, for various reasons).
 * Added additional APIs for obtaining new collision-free enum case values.
 * Actually implemented part-type-exclusive artifacts (these used to do nothing).
+* `[ModBuildConfig 1.0.0]` Legacy mods can now use .NET 8 and interact with Nickel directly via a new `INickelManifest` interface.
 
 ## 0.12.0
 Released 15 May 2024.